refactor(util): group room/zoom-room helpers under util/rooms/
Moves rooms, filterColors, zoomHosts, and modeIcons into util/rooms/ --
four files of room/zoom-room lookup tables and option lists that
previously sat flat alongside unrelated util files. Pure relocation.
