+++ frontend/app/components/admin/diagnostics/ConflictList.tsx
@@ -4,7 +4,7 @@ import React, { useMemo, useState } from "react";
 import EditMeetingSidebar from "../../meeting-form/EditMeeting";
 import { IMeeting } from "../../../../types/models";
 import { useZoomHostPool } from "../../../../hooks/useZoomHostPool";
-import { zoomHostLabel } from "../../../../util/zoomHosts";
+import { zoomHostLabel } from "../../../../util/rooms/zoomHosts";
 import {
   ConflictListRow,
   formatOverlapSummary,
