+++ frontend/app/components/calendar/mobile/DayLandscapeView.tsx
@@ -6,7 +6,7 @@ import { fetchMeetingsByDay, invalidateCache } from "../desktop/DayView";
 import { formatETDateString, getCurrentETMinutesSinceMidnight } from "../../../../util/date/timeUtils";
 import { formatMeetingDateLine } from "../../../../util/date/timeFormat";
 import { passesTagFilters, passesRoomFilter, MeetingFilters } from "../../../../util/meetingFilters";
-import { defaultRooms } from "../../../../util/rooms";
+import { defaultRooms } from "../../../../util/rooms/rooms";
 import { layoutOverlappingMeetings, OverlapMeeting } from "../../../../util/meetingOverlapLayout";
 import { addDaysToDate } from "../../../../util/date/weekDates";
 import { useElementSize } from "../../../../hooks/useElementSize";
