+++ frontend/app/components/calendar/shared/DayColumn.tsx
@@ -2,7 +2,7 @@ import React, { useState, useRef, useEffect } from 'react';
 import BoxText from '../../atoms/BoxText';
 import OverlapMeetingsModal from './OverlapMeetingsModal';
 import styles from '../../../../styles/components/calendar/shared/DayColumn.module.scss';
-import { isZoomRoomMismatched } from '../../../../util/rooms';
+import { isZoomRoomMismatched } from '../../../../util/rooms/rooms';
 import { formatCompactTimeRange } from '../../../../util/date/timeFormat';
 import { formatETDateString } from '../../../../util/date/timeUtils';
