+++ frontend/app/components/meeting-form/ViewMeeting.tsx
@@ -14,12 +14,12 @@ import { formatCompactTimeRange, formatMeetingDateLine } from "../../../util/dat
 import { formatETDateString } from "../../../util/date/timeUtils";
 import { retryMeetingSync } from "../../../services/syncMeeting";
 import { formatSuspensionStatusText } from "../../../util/suspensionText";
-import { ROOM_COLORS, ZOOM_ROOM_COLOR } from "../../../util/filterColors";
+import { ROOM_COLORS, ZOOM_ROOM_COLOR } from "../../../util/rooms/filterColors";
 import { formatRecurrencePattern } from "../../../util/recurrenceDisplay";
-import { isZoomRoomMismatched } from "../../../util/rooms";
+import { isZoomRoomMismatched } from "../../../util/rooms/rooms";
 import { linkify } from "../../../util/common/linkify";
-import { zoomHostLabel } from "../../../util/zoomHosts";
-import { MODE_ICON_SRC } from "../../../util/modeIcons";
+import { zoomHostLabel } from "../../../util/rooms/zoomHosts";
+import { MODE_ICON_SRC } from "../../../util/rooms/modeIcons";
 import { useZoomHostPool } from "../../../hooks/useZoomHostPool";
 
 // Extracts ET wall-clock time as "HH:MM" (24hr), which is what formatCompactTimeRange expects.
