+++ frontend/hooks/useMeetingForm.ts
@@ -1,7 +1,7 @@
 import { useCallback, useState } from 'react';
 import { IMeeting, IRecurrencePattern } from '../types/models';
 import { convertUTCToET, convertETToUTC, formatETDateString, getWeekDatesET, getCurrentETMinutesSinceMidnight } from '../util/date/timeUtils';
-import { roomToZoomRoom } from '../util/rooms';
+import { roomToZoomRoom } from '../util/rooms/rooms';
 import { DESCRIPTION_MAX_LENGTH } from '../util/meetingValidation';
 
 // What the calendar is currently showing, used to seed a brand-new meeting's default
