+++ frontend/tests/component/DayLandscapeView.test.tsx
@@ -3,7 +3,7 @@ import { render, within } from "@testing-library/react";
 import DayLandscapeView, { roomDisplayName } from "../../app/components/calendar/mobile/DayLandscapeView";
 import { CalendarProvider } from "../../app/context/CalendarProvider";
 import { createDefaultFilters } from "../../util/meetingFilters";
-import { defaultRooms } from "../../util/rooms";
+import { defaultRooms } from "../../util/rooms/rooms";
 import { formatETDateString } from "../../util/date/timeUtils";
 
 // Noon ET (UTC-4, matching this repo's other fixed-date component tests, e.g.
