+++ frontend/util/rooms/modeIcons.ts
